MLOps definition - add examples for add't clarity
I agree with the statement /*The extension of the DevOps methodology to include Machine Learning and Data Science Assets and the interdependencies between those Assets as first class citizens within the DevOps ecology.*/

I added some specific examples to help increase the clarity of what is needed.  I tried to create a high level list of assets that was representative, but would welcome feedback if there are additional assets worth calling out or if we should go to the next level of depth.